work on ttt and fix of w2p1

+++ W2P1/main.cpp
@@ -14,7 +14,7 @@ int main() {
         float sum;
         cin >> input;
         if (input == -1) {
-            sum = sum / inputCount;
+            sum = sum / (float)inputCount;
             cout << "average = " << sum << endl;
             break;
         }
